// projects/investbetz-platform/src/app/client/feedback/testimonial/testimonial.interface.ts
export interface TestimonialInterface {
    clientId: string,
    situationBefore: string;
    resultsFromService: string;
    coreProblem: string;
    whatHesitation: string;
    specificFeatures: string;
    recommendation: string;
    anythingElse: string;
    testimonialCheck: boolean;
    //testimonialDate: Date;
}